# server/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/posts/upload": {
      "post": {
        "summary": "Upload Post",
        "tags": ["Posts"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post uploaded successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UploadPost": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}